Implement retransmission of data when endpoints connect

+++ runtime/nrf52/platform/nrf52_print.c
@@ -8,11 +8,11 @@
 void
 ravel_system_print(RavelSystemAPI *self, const char *msg)
 {
-    NRF_LOG_DEBUG("%s\r\n",msg);
+    NRF_LOG_DEBUG("%s\r\n", (uint32_t)msg);
 }
 
 void
 ravel_system_print_number(RavelSystemAPI *self, const char *msg, int32_t number)
 {
-    NRF_LOG_DEBUG("%s %d\r\n",msg, number);
+    NRF_LOG_DEBUG("%s %d\r\n", (uint32_t)msg, number);
 }
